+++ hydra-node/test/Hydra/API/ServerSpec.hs
@@ -7,7 +7,6 @@ import Hydra.Prelude hiding (decodeUtf8, seq)
 import Test.Hydra.Prelude
 
 import Cardano.Binary (serialize')
-import Control.Lens ((^?))
 import Control.Concurrent.Class.MonadSTM (
   check,
   modifyTVar',
@@ -18,10 +17,11 @@ import Control.Concurrent.Class.MonadSTM (
   tryReadTQueue,
   writeTQueue,
  )
+import Control.Lens ((^?))
 import qualified Data.Aeson as Aeson
-import qualified Data.Aeson.KeyMap as KeyMap
 import Data.Aeson.Lens (key, nonNull)
 import qualified Data.ByteString.Base16 as Base16
+import qualified Data.List as List
 import qualified Data.Text as T
 import Data.Text.Encoding (decodeUtf8)
 import Hydra.API.Server (RunServerException (..), Server (Server, sendOutput), withAPIServer)
@@ -31,7 +31,7 @@ import Hydra.Ledger.Simple (SimpleTx)
 import Hydra.Logging (showLogsOnFailure)
 import Hydra.Network (PortNumber)
 import Hydra.Persistence (PersistenceIncremental (..), createPersistenceIncremental)
-import Hydra.Snapshot (ConfirmedSnapshot (..), Snapshot, confirmed)
+import Hydra.Snapshot (ConfirmedSnapshot (..), Snapshot (Snapshot, utxo), confirmed)
 import Network.WebSockets (Connection, receiveData, runClient, sendBinaryData)
 import System.IO.Error (isAlreadyInUseError)
 import System.Timeout (timeout)
@@ -60,10 +60,7 @@ spec = describe "ServerSpec" $
           withFreePort $ \port ->
             withAPIServer @SimpleTx "127.0.0.1" port alice mockPersistence tracer noop $ \_ -> do
               withClient port "/" $ \conn -> do
-                received <- receiveData conn
-                case Aeson.eitherDecode received of
-                  Left{} -> failure $ "Failed to decode greeting " <> show received
-                  Right TimedServerOutput{output = msg} -> msg `shouldBe` greeting
+                waitMatch 5 conn $ guard . matchGreetings
 
     it "sends sendOutput to all connected clients" $ do
       queue <- atomically newTQueue
@@ -78,7 +75,10 @@ spec = describe "ServerSpec" $
               )
               $ \_ -> do
                 waitForClients semaphore
-                failAfter 1 $ atomically (replicateM 2 (readTQueue queue)) `shouldReturn` [greeting, greeting]
+                failAfter 1 $
+                  atomically (replicateM 2 (readTQueue queue))
+                    >>= (`shouldSatisfyAll` [isGreetings, isGreetings])
+
                 arbitraryMsg <- generate arbitrary
                 sendOutput arbitraryMsg
                 failAfter 1 $ atomically (replicateM 2 (readTQueue queue)) `shouldReturn` [arbitraryMsg, arbitraryMsg]
@@ -108,12 +108,23 @@ spec = describe "ServerSpec" $
                 )
                 $ \_ -> do
                   waitForClients semaphore
-                  failAfter 1 $ atomically (replicateM 2 (readTQueue queue1)) `shouldReturn` [arbitraryMsg, greeting]
-                  failAfter 1 $ atomically (replicateM 2 (readTQueue queue2)) `shouldReturn` [arbitraryMsg, greeting]
+                  failAfter 1 $
+                    atomically (replicateM 2 (readTQueue queue1))
+                      >>= flip shouldSatisfyAll [(==) arbitraryMsg, isGreetings]
+                  failAfter 1 $
+                    atomically (replicateM 2 (readTQueue queue2))
+                      >>= flip shouldSatisfyAll [(==) arbitraryMsg, isGreetings]
+
                   sendOutput arbitraryMsg
-                  failAfter 1 $ atomically (replicateM 1 (readTQueue queue1)) `shouldReturn` [arbitraryMsg]
-                  failAfter 1 $ atomically (replicateM 1 (readTQueue queue2)) `shouldReturn` [arbitraryMsg]
-                  failAfter 1 $ atomically (tryReadTQueue queue1) `shouldReturn` Nothing
+                  failAfter 1 $
+                    atomically (replicateM 1 (readTQueue queue1))
+                      `shouldReturn` [arbitraryMsg]
+                  failAfter 1 $
+                    atomically (replicateM 1 (readTQueue queue2))
+                      `shouldReturn` [arbitraryMsg]
+                  failAfter 1 $
+                    atomically (tryReadTQueue queue1)
+                      `shouldReturn` Nothing
 
     it "echoes history (past outputs) to client upon reconnection" $
       checkCoverage . monadicIO $ do
@@ -131,7 +142,9 @@ spec = describe "ServerSpec" $
                   case traverse Aeson.eitherDecode received of
                     Left{} -> failure $ "Failed to decode messages:\n" <> show received
                     Right timedOutputs -> do
-                      (output <$> timedOutputs) `shouldBe` outputs <> [greeting]
+                      let actualOutputs = output <$> timedOutputs
+                      List.init actualOutputs `shouldBe` outputs
+                      List.last actualOutputs `shouldSatisfy` isGreetings
 
     it "does not echo history if client says no" $
       checkCoverage . monadicIO $ do
@@ -148,11 +161,7 @@ spec = describe "ServerSpec" $
                 -- start client that doesn't want to see the history
                 withClient port "/?history=no" $ \conn -> do
                   -- wait on the greeting message
-                  waitMatch 5 conn $ \v ->
-                    case Aeson.fromJSON v of
-                      Aeson.Success timedOutput ->
-                        guard $ output timedOutput == greeting
-                      _other -> Nothing
+                  waitMatch 5 conn $ guard . matchGreetings
 
                   notHistoryMessage :: ServerOutput SimpleTx <- generate arbitrary
                   sendFromApiServer notHistoryMessage
@@ -196,10 +205,7 @@ spec = describe "ServerSpec" $
                     , postTxError = NoSeedInput
                     }
                 guardForValue v expected =
-                  case v of
-                    Aeson.Object km ->
-                      guard . (expected ==) =<< KeyMap.lookup "transaction" km
-                    _other -> Nothing
+                  guard $ v ^? key "transaction" == Just expected
 
             -- client is able to specify they want tx output to be encoded as CBOR
             withClient port "/?history=no&tx-output=cbor" $ \conn -> do
@@ -238,18 +244,18 @@ spec = describe "ServerSpec" $
         withFreePort $ \port ->
           withAPIServer @SimpleTx "127.0.0.1" port alice mockPersistence tracer noop $ \Server{sendOutput} -> do
             snapshot <- generate arbitrary
-            let snapshotConfirmedMessage = SnapshotConfirmed{headId = HeadId "some-head-id", Hydra.API.ServerOutput.snapshot, Hydra.API.ServerOutput.signatures = mempty}
+            let snapshotConfirmedMessage =
+                  SnapshotConfirmed
+                    { headId = HeadId "some-head-id"
+                    , Hydra.API.ServerOutput.snapshot
+                    , Hydra.API.ServerOutput.signatures = mempty
+                    }
 
             withClient port "/?snapshot-utxo=no" $ \conn -> do
               sendOutput snapshotConfirmedMessage
 
               waitMatch 5 conn $ \v ->
-                case v of
-                  Aeson.Object km -> do
-                    case KeyMap.lookup "snapshot" km of
-                      Just (Aeson.Object km') -> guard $ isNothing $ KeyMap.lookup "utxo" km'
-                      _other -> Nothing
-                  _other -> Nothing
+                guard $ isNothing $ v ^? key "utxo"
 
     it "sequence numbers are continuous and strictly monotonically increasing" $
       monadicIO $ do
@@ -267,9 +273,67 @@ spec = describe "ServerSpec" $
                     Right (timedOutputs :: [TimedServerOutput SimpleTx]) ->
                       seq <$> timedOutputs `shouldSatisfy` strictlyMonotonic
 
+    it "displays correctly headStatus and snapshotUtxo in a Greeting message" $
+      showLogsOnFailure $ \tracer ->
+        withFreePort $ \port -> do
+          withAPIServer @SimpleTx "127.0.0.1" port alice mockPersistence tracer noop $ \Server{sendOutput} -> do
+            let generateSnapshot =
+                  generate $
+                    SnapshotConfirmed <$> arbitrary <*> arbitrary <*> arbitrary
+
+            waitForValue port $ \v -> do
+              guard $ v ^? key "headStatus" == Just (Aeson.String "Idle")
+              -- test that the 'snapshotUtxo' is excluded from json if there is no utxo
+              guard $ isNothing (v ^? key "snapshotUtxo")
+
+            headIsOpenMsg <- generate $ HeadIsOpen <$> arbitrary <*> arbitrary
+            snapShotConfirmedMsg@SnapshotConfirmed{snapshot = Snapshot{utxo}} <-
+              generateSnapshot
+
+            mapM_ sendOutput [headIsOpenMsg, snapShotConfirmedMsg]
+            waitForValue port $ \v -> do
+              guard $ v ^? key "headStatus" == Just (Aeson.String "Open")
+              guard $ v ^? key "snapshotUtxo" == Just (toJSON utxo)
+
+            snapShotConfirmedMsg'@SnapshotConfirmed{snapshot = Snapshot{utxo = utxo'}} <-
+              generateSnapshot
+            let readyToFanoutMsg = ReadyToFanout $ headId headIsOpenMsg
+
+            mapM_ sendOutput [readyToFanoutMsg, snapShotConfirmedMsg']
+            waitForValue port $ \v -> do
+              guard $ v ^? key "headStatus" == Just (Aeson.String "FanoutPossible")
+              guard $ v ^? key "snapshotUtxo" == Just (toJSON utxo')
+
+    it "greets with correct head status and snapshot utxo after restart" $
+      showLogsOnFailure $ \tracer ->
+        withTempDir "api-server-head-status" $ \persistenceDir ->
+          withFreePort $ \port -> do
+            let generateSnapshot =
+                  generate $
+                    SnapshotConfirmed <$> arbitrary <*> arbitrary <*> arbitrary
+            apiPersistence <- createPersistenceIncremental $ persistenceDir <> "/server-output"
+            snapShotConfirmedMsg@SnapshotConfirmed{snapshot = Snapshot{utxo}} <-
+              generateSnapshot
+            let expectedUtxos = toJSON utxo
+
+            withAPIServer @SimpleTx "127.0.0.1" port alice apiPersistence tracer noop $ \Server{sendOutput} -> do
+              headIsInitializing <- generate $ HeadIsInitializing <$> arbitrary <*> arbitrary
+
+              mapM_ sendOutput [headIsInitializing, snapShotConfirmedMsg]
+              waitForValue port $ \v -> do
+                guard $ v ^? key "headStatus" == Just (Aeson.String "Initializing")
+                guard $ v ^? key "snapshotUtxo" == Just expectedUtxos
+
+            -- expect the api server to load events from apiPersistence and project headStatus correctly
+            withAPIServer @SimpleTx "127.0.0.1" port alice apiPersistence tracer noop $ \_ -> do
+              waitForValue port $ \v -> do
+                guard $ v ^? key "headStatus" == Just (Aeson.String "Initializing")
+                guard $ v ^? key "snapshotUtxo" == Just expectedUtxos
+
     it "sends an error when input cannot be decoded" $
       failAfter 5 $
-        withFreePort $ \port -> sendsAnErrorWhenInputCannotBeDecoded port
+        withFreePort $
+          \port -> sendsAnErrorWhenInputCannotBeDecoded port
 
 strictlyMonotonic :: [Natural] -> Bool
 strictlyMonotonic = \case
@@ -294,8 +358,14 @@ sendsAnErrorWhenInputCannotBeDecoded port = do
     InvalidInput{input} -> input == invalidInput
     _ -> False
 
-greeting :: ServerOutput SimpleTx
-greeting = Greetings alice
+matchGreetings :: Aeson.Value -> Bool
+matchGreetings v =
+  v ^? key "tag" == Just (Aeson.String "Greetings")
+
+isGreetings :: ServerOutput tx -> Bool
+isGreetings = \case
+  Greetings{} -> True
+  _ -> False
 
 waitForClients :: (MonadSTM m, Ord a, Num a) => TVar m a -> m ()
 waitForClients semaphore = atomically $ readTVar semaphore >>= \n -> check (n >= 2)
@@ -327,6 +397,11 @@ mockPersistence =
     , loadAll = pure []
     }
 
+waitForValue :: PortNumber -> (Aeson.Value -> Maybe ()) -> IO ()
+waitForValue port f =
+  withClient port "/?history=no" $ \conn ->
+    waitMatch 5 conn f
+
 -- | Wait up to some time for an API server output to match the given predicate.
 waitMatch :: HasCallStack => Natural -> Connection -> (Aeson.Value -> Maybe a) -> IO a
 waitMatch delay con match = do
@@ -356,3 +431,14 @@ waitMatch delay con match = do
     case Aeson.eitherDecode' bytes of
       Left err -> failure $ "WaitNext failed to decode msg: " <> err
       Right value -> pure value
+
+shouldSatisfyAll :: Show a => [a] -> [a -> Bool] -> Expectation
+shouldSatisfyAll values predicates =
+  go values predicates
+ where
+  go [] [] = pure ()
+  go [] _ = failure "shouldSatisfyAll: ran out of values"
+  go _ [] = failure "shouldSatisfyAll: ran out of predicates"
+  go (v : vs) (p : ps) = do
+    v `shouldSatisfy` p
+    go vs ps